fix(weapp-vite): normalize copied miniprogram esm modules
